% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{app_indicators}
\alias{app_indicators}
\title{Components relating to the economic indicators module}
\usage{
app_indicators()
}
\description{
This function provides the components for running the economic indicators module, be that as a sub-module of a bigger app or on a standalone basis.
}
\examples{
\dontrun{
app_indicators()
}

}
\keyword{module}
