% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{app_main}
\alias{app_main}
\title{Running the main app that controls the shiny sub modules}
\usage{
app_main()
}
\description{
This function runs the main app that controls the shiny sub-modules.
}
\examples{
\dontrun{
app_main()
}
}
\keyword{module}
