% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectors.R
\name{app_sectors}
\alias{app_sectors}
\title{Components relating to the sector performance module}
\usage{
app_sectors()
}
\description{
This function provides the components for running the sector performance module, be that as a sub-module of a bigger app or on a standalone basis.
}
\examples{
\dontrun{
app_sectors()
}
}
\keyword{module}
