% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_indicators.R
\name{get_data_indicators}
\alias{get_data_indicators}
\title{Imports data relating to key US economic indicators over time}
\usage{
get_data_indicators()
}
\description{
This function enables you to import data relating to key US economic indicators over time.
}
\examples{
\dontrun{
get_data_indicators()
}
}
\keyword{data}
