% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_sectors.R
\name{get_data_sectors}
\alias{get_data_sectors}
\title{Imports data relating to the performance of US sectors over time}
\usage{
get_data_sectors()
}
\description{
This function enables you to import data relating to the performance of US sectors over time.
}
\examples{
\dontrun{
get_data_sectors()
}
}
\keyword{data}
