% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{server_indicators}
\alias{server_indicators}
\title{Components relating to the economic indicators module}
\usage{
server_indicators(id, data_import)
}
\arguments{
\item{id}{The identifier used in shiny to namespace a module, thereby linking the ui and server components}

\item{data_import}{The data imported on economic indicators}
}
\description{
This function provides the components for running the economic indicators module, be that as a sub-module of a bigger app or on a standalone basis.
}
\examples{
\dontrun{
server_indicators(id = "indicators", data_import = data_imported)
}

}
\keyword{module}
