% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectors.R
\name{server_sectors}
\alias{server_sectors}
\title{Components relating to the sector performance module}
\usage{
server_sectors(id, data_import)
}
\arguments{
\item{id}{The identifier used in shiny to namespace a module, thereby linking the ui and server components}

\item{data_import}{The data imported on sector performance}
}
\description{
This function provides the components for running the sector performance module, be that as a sub-module of a bigger app or on a standalone basis.
}
\examples{
\dontrun{
server_sectors(id = "sectors", data_import = data_imported)
}

}
\keyword{module}
