% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ui_indicators}
\alias{ui_indicators}
\title{Components relating to the economic indicators module}
\usage{
ui_indicators(id, standalone = FALSE)
}
\arguments{
\item{id}{The identifier used in shiny to namespace a module, thereby linking the ui and server components}

\item{standalone}{Whether to view the module on a standalone basis (TRUE) or not (FALSE, the default)}

\item{data_import}{The data imported on economic indicators}
}
\description{
This function provides the components for running the economic indicators module, be that as a sub-module of a bigger app or on a standalone basis.
}
\examples{
\dontrun{
app_indicators(standalone = TRUE)
}
}
\keyword{module}
